requirements.txt
dbutils==3.1.2
Flask==3.1.2
flask_cors==6.0.1
locustio==0.999
pymysql==1.1.2
redis==7.1.0
waitress==3.0.2
